Move stuff from PuddleSystem -> SpilableSystem
